+++ lib/buildkite/test_collector/uuid.rb
@@ -4,9 +4,14 @@ require "securerandom"
 
 class Buildkite::TestCollector::UUID
   GET_UUID = SecureRandom.method(:uuid)
-  private_constant :GET_UUID
+  GET_UUID_V7 = SecureRandom.method(:uuid_v7)
+  private_constant :GET_UUID, :GET_UUID_V7
 
   def self.call
     GET_UUID.call
+  end
+
+  def self.v7
+    GET_UUID_V7.call
   end
 end